requirements.txt
ansi2html==1.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
category-encoders==2.5.1.post0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
cramjam==2.6.2
Cython==3.0.0a11
dash==2.7.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fastparquet==2022.12.0
Flask==2.1.3
Flask-Cors==3.0.10
fqdn==1.5.1
fsspec==2022.11.0
idna==3.4
imbalanced-learn==0.10.0
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter-dash==0.4.2
jupyter-events==0.5.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
jupyterlab-widgets==3.0.4
lightgbm==3.3.3
llvmlite==0.38.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.55.2
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pip==22.3.1
platformdirs==2.6.0
plotly-resampler==0.8.3
pmdarima==1.8.5
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pycaret==3.0.0rc4
pycparser==2.21
Pygments==2.13.0
pyod==1.0.6
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.8.1
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
sktime==0.13.4
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
tbats==1.1.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
wheel==0.37.1
wincertstore==0.2
yellowbrick==1.5
